ai_pc_agent/requirements.txt
# ── Speech Recognition ────────────────────────────────────────────────────────
speechrecognition>=3.10.0
pyaudio>=0.2.13

# ── Automation ────────────────────────────────────────────────────────────────
pyautogui>=0.9.54
keyboard>=0.13.5
pyperclip>=1.8.2

# ── System ────────────────────────────────────────────────────────────────────
psutil>=5.9.0
pygetwindow>=0.0.9

# ── TTS ───────────────────────────────────────────────────────────────────────
pyttsx3>=2.90

# ── HTTP / LLM ────────────────────────────────────────────────────────────────
requests>=2.31.0

# ── Vision / Screenshot ───────────────────────────────────────────────────────
numpy>=1.24.0
opencv-python>=4.8.0
pillow>=10.0.0
mss>=9.0.1

# ── Console / Logging ─────────────────────────────────────────────────────────
rich>=13.7.0
